+++ core/src/index/thirdparty/faiss/gpu/GpuIndexIVFFlat.h
@@ -70,7 +70,8 @@ class GpuIndexIVFFlat : public GpuIndexIVF {
                    const float* x,
                    int k,
                    float* distances,
-                   Index::idx_t* labels) const override;
+                   Index::idx_t* labels,
+                   ConcurrentBitsetPtr bitset = nullptr) const override;
 
  private:
   GpuIndexIVFFlatConfig ivfFlatConfig_;